Encoded names (#30)
* admin: add validatins check to the selected files before upload

* server: add encoded name and path to the fileData field in the layer shcema

* server: check upload

* server: return files without the unuseful information. cancel the removing of the zip directory from the upload temp directory

* admin: add the reqFile fields inside a file objcet

* server: add encoded names and checks to the upload vector's files

* admin: add the encoded names + a message if the upload failed

* server: add the public directory to the git ignore file

* server: add the json directory to the git ignore file

* tb: add the public directory to the git ignore file

* server: add file system router for removing files and directories by request

* admin: add - removing the vector zip directory after the first time the displaying the map

+++ server/src/api/geoserverCrud/UploadFilesToGS.js
@@ -1,43 +1,36 @@
-require('../fileMethods')();
+// const GsLayers  = require("../geoserverCrud/GsLayers");
+require('../fs/fileMethods')();
 require('./curlMethods')();
+// require('../../config/serverConfig')();
+// const configParams = config().configParams;
+// const configUrl = configBaseUrl().configUrl;
 
 class UploadFilesToGS {
 
     static uploadFile(workspaceName, reqFiles, name, path) {
-    		const file = reqFiles.length ? reqFiles : [reqFiles];
-				console.log("uploadFile file: " + JSON.stringify(file));
+    		let files = reqFiles.length ? reqFiles : [reqFiles];
+				console.log("uploadFile files: " + JSON.stringify(files));
 				console.log("uploadFile PATH: " + path);
 
-        // 1. create the JSON file with the desire workspace
+        // 1. create the JSON files with the desire workspace
         let importJSON = {};
 
-				console.log("file Type: " + file[0].fileType);
-        if (file[0].fileType.toLowerCase() === 'raster') {
+				console.log("files Type: " + files[0].fileType);
+        if (files[0].fileType.toLowerCase() === 'raster') {
             importJSON = createImportObject(workspaceName);
         } else {
             importJSON = createImportObjectWithData(workspaceName, path);
         }
         console.log("importJSON: " + JSON.stringify(importJSON));
 
-        // 2. get the Import object by POST the JSON file and check it
+        // 2. get the Import object by POST the JSON files and check it
         let importObj = postImportObj(JSON.stringify(importJSON));
         console.log("import: " + JSON.stringify(importObj));
 
         // 3a. for VECTORS only:
-				if (file[0].fileType.toLowerCase() === 'vector') {
-					// A. check the FORMAT field in the data field in the Import Object and update it if it's missing
-					if(!importObj.data.format || importObj.data.format == null){
-						importObj.data.format = 'Shapefile';
-						const updateImportJson = JSON.stringify({ import: importObj });
-						console.log("NO_FORMAT updateImportJson = " + updateImportJson);
-					}
-
-					// B. get the Task List from the Import Object
-					// let taskList = importObj.tasks;
-					// console.log("task List: " + JSON.stringify(taskList));
-
-					// C. check the STATE of each task in the Task List
-					console.log("taskList map... ");
+				if (files[0].fileType.toLowerCase() === 'vector') {
+					// check the STATE of each task in the Task List
+					console.log("check the state of each task... ");
 					importObj.tasks.map( task => {
 						console.log(`task ${task.id} state = ${task.state}`);
 						if (task.state !== 'READY') {
@@ -45,41 +38,18 @@ class UploadFilesToGS {
 							task = getTaskObj(importObj.id, task.id);
 							console.log(`task ${task.id} (before change): ${JSON.stringify(task)}`);
 							// check the state's error and fix it
-							switch (task.state) {
-								case ('NO_FORMAT'):
-									console.log("NO_FORMAT updateTaskJson: ");
-									// create the update FORMAT Json file and update the import's data
-									// updateTaskJson = JSON.stringify(dataFormatUpdate());
-									task.data.format = 'Shapefile';
-									task.transformChain.type = 'vector';
-									const updateTaskJson = JSON.stringify({ task: task });
-									console.log(`task ${task.id} (after the change): ${updateTaskJson}`);
-									updateTaskById(updateTaskJson, importObj.id, task.id);
-									break;
-								case ('NO_CRS'):
-									// get the layer Object from the Task Object and update the projection
-									// task.layer = getLayerObj(importObj.id, task.id);
-									// task.layer.srs = 'EPSG:4326';
-									// const updateLayerJson = JSON.stringify({ layer: task.layer });
-									const updateLayerJson = JSON.stringify(layerSrsUpdate());
-									console.log("NO_CRS updateTaskJson: " + updateLayerJson);
-									// create the update SRS Json file and update the task
-									updateTaskField(updateLayerJson, importObj.id, task.id, 'layer');
-									break;
-								default:
-									// return an Error Massage with the task's state
-									return `ERROR to upload the file! state: ${task.state}`;
+							if (task.state === 'NO_CRS') {
+								const updateLayerJson = JSON.stringify(layerSrsUpdate());
+								console.log("NO_CRS updateTaskJson: " + updateLayerJson);
+								// create the update SRS Json files and update the task
+								updateTaskField(updateLayerJson, importObj.id, task.id, 'layer');
+							} else {
+									files = [];
 							}
 						}
 					});
-					// console.log("update import(from the object): " + JSON.stringify(importObj));
-					// D. update the Import Object
-					// updateImportById(importObj, importObj.id);
-					// get the update Import Object
-					// importObj = getImportObj(importObj.id);
-					// console.log("update import(from geoserver): " + JSON.stringify(importObj));
 				}
-				// 3b. for RASTERS only: POST the file to the tasks list, in order to create an import task for it
+				// 3b. for RASTERS only: POST the files to the tasks list, in order to create an import task for it
 				else {
 					sendToTask(path, name, importObj.id);
 				}
@@ -87,19 +57,32 @@ class UploadFilesToGS {
         // 4. execute the import task
         executeFileToGeoserver(importObj.id);
 
-        // 5. remove the file from the local store
-        removeFile(path);
-				// in ZIP file: remove the directory (after open the zip file)
+				// 5. delete all the uncompleted tasks in the import queue
+				deleteUncompleteImports();
+
+				// 6. remove the files from the local store
+				removeFile(path);
+				// if ZIP files:
+				// send the path in the return files object to remove the zip directory after uploading the layer in geoserver
 				const splitPath = path.split('.');
-        if (splitPath[1] === 'zip'){
-					removeFile(splitPath[0]);
+				if (splitPath[1] === 'zip'){
+					files.map( file => {
+						if (file.fileType.toLowerCase() === 'vector'){
+							file.splitPath = splitPath[0].trim();
+						} else {
+							file.splitPath = null;
+							removeFile(splitPath[0]);
+						}
+					});
+				} else {
+					console.log("the files is no a ZIP!");
+					files[0].splitPath = null;
+					console.log("splitPath: " + files[0].splitPath);
 				}
 
-        // 6. delete all the uncompleted tasks in the import queue
-        deleteUncompleteImports();
-
-        // 7. return OK
-        return file;
+				// 7. return the files
+				console.log("return files: " + JSON.stringify(files));
+				return files;
     }
 }
